reducers modilarized
