+++ src/actions/index.js
@@ -1,7 +1,8 @@
 // Action types:
-export const UPDATE_ITEMS = 'UPDATE_ITEMS';
-export const ADD_ITEM = 'ADD_ITEM';
+export const ADD_ITEM    = 'ADD_ITEM';
 export const DELETE_ITEM = 'DELETE_ITEM';
+export const SEARCH_GIFS = 'SEARCH_GIFS';
+export const ALL_GIFS    = 'ALL_GIFS';
 
 //Action creators
 export const addItem = (payload) => ({
@@ -14,8 +15,12 @@ export const deleteItem = (payload) => ({
     payload
 })
 
-export const updateItems = (payload) => ({
-    type : UPDATE_ITEMS,
+export const searchGifs = (payload) => ({
+    type : SEARCH_GIFS,
+    payload
+})
+export const allGifs = (payload) => ({
+    type : All_GIFS,
     payload
 })
